# packages/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: EdgeReco API
  version: "0.0.0"
  description: Edge recommendation engine — Phase 0 proof-of-loop API

servers:
  - url: http://localhost:8000

paths:
  /v0/healthz:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    enum:
                      - ok

  /v0/catalog:
    get:
      operationId: getCatalog
      summary: Retrieve the full item catalog
      responses:
        "200":
          description: Full catalog with generation timestamp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogResponse"

  /v0/candidates:
    post:
      operationId: getCandidates
      summary: Request candidate items for reranking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateRequest"
      responses:
        "200":
          description: Candidate items matching the request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateResponse"

  /v0/events:
    post:
      operationId: ingestEvents
      summary: Ingest a batch of interaction events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventBatch"
      responses:
        "202":
          description: Events accepted
          content:
            application/json:
              schema:
                type: object
                required:
                  - received
                properties:
                  received:
                    type: integer

components:
  schemas:
    CatalogItem:
      type: object
      required:
        - id
        - title
        - category
        - tags
        - popularityScore
        - freshnessScore
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        popularityScore:
          type: number
          minimum: 0
          maximum: 1
        freshnessScore:
          type: number
          minimum: 0
          maximum: 1

    CandidateRequest:
      type: object
      required:
        - contextType
        - limit
      properties:
        contextType:
          type: string
        categoryHint:
          type: string
        limit:
          type: integer
          minimum: 1
          maximum: 200

    CandidateResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"

    CatalogResponse:
      type: object
      required:
        - items
        - generatedAt
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
        generatedAt:
          type: string
          format: date-time

    Event:
      type: object
      required:
        - eventId
        - eventType
        - itemId
        - timestamp
        - contextType
      properties:
        eventId:
          type: string
        eventType:
          type: string
          enum:
            - impression
            - click
            - favorite
        itemId:
          type: string
        timestamp:
          type: string
          format: date-time
        contextType:
          type: string

    EventBatch:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"