# services/api/app/generated/models.py
# AUTO-GENERATED by packages/contracts/scripts/generate-python.sh
# DO NOT EDIT BY HAND. Regenerate with `pnpm -C packages/contracts generate:py`.

# generated by datamodel-codegen:
#   filename:  openapi.yaml

from __future__ import annotations

from datetime import datetime
from enum import Enum
from typing import List, Optional

from pydantic import BaseModel, Field


class CatalogItem(BaseModel):
    id: str
    title: str
    category: str
    tags: List[str]
    popularity_score: float = Field(..., alias='popularityScore', ge=0.0, le=1.0)
    freshness_score: float = Field(..., alias='freshnessScore', ge=0.0, le=1.0)


class CandidateRequest(BaseModel):
    context_type: str = Field(..., alias='contextType')
    category_hint: Optional[str] = Field(None, alias='categoryHint')
    limit: int = Field(..., ge=1, le=200)


class CandidateResponse(BaseModel):
    items: List[CatalogItem]


class CatalogResponse(BaseModel):
    items: List[CatalogItem]
    generated_at: datetime = Field(..., alias='generatedAt')


class EventType(Enum):
    impression = 'impression'
    click = 'click'
    favorite = 'favorite'


class Event(BaseModel):
    event_id: str = Field(..., alias='eventId')
    event_type: EventType = Field(..., alias='eventType')
    item_id: str = Field(..., alias='itemId')
    timestamp: datetime
    context_type: str = Field(..., alias='contextType')


class EventBatch(BaseModel):
    events: List[Event]


class Status(Enum):
    ok = 'ok'


class V0EventsPostResponse(BaseModel):
    received: int


class V0HealthzGetResponse(BaseModel):
    status: Status