# services/api/app/generated/routes.py
# AUTO-GENERATED by packages/contracts/scripts/generate-python.sh
# DO NOT EDIT BY HAND. Regenerate with `pnpm -C packages/contracts generate:py`.

# generated by fastapi-codegen:
#   filename:  openapi.yaml

from __future__ import annotations

from typing import Optional, Union

from fastapi import FastAPI

from .models import (
    CandidateRequest,
    CandidateResponse,
    CatalogResponse,
    EventBatch,
    V0EventsPostResponse,
    V0HealthzGetResponse,
)

app = FastAPI(
    title='EdgeReco API',
    version='0.0.0',
    description='Edge recommendation engine — Phase 0 proof-of-loop API',
    servers=[{'url': 'http://localhost:8000'}],
)


@app.post('/v0/candidates', response_model=CandidateResponse)
def get_candidates(body: CandidateRequest) -> CandidateResponse:
    """
    Request candidate items for reranking
    """
    pass


@app.get('/v0/catalog', response_model=CatalogResponse)
def get_catalog() -> CatalogResponse:
    """
    Retrieve the full item catalog
    """
    pass


@app.post(
    '/v0/events',
    response_model=None,
    responses={'202': {'model': V0EventsPostResponse}},
)
def ingest_events(body: EventBatch) -> Optional[V0EventsPostResponse]:
    """
    Ingest a batch of interaction events
    """
    pass


@app.get('/v0/healthz', response_model=V0HealthzGetResponse)
def get_health() -> V0HealthzGetResponse:
    """
    Health check
    """
    pass